# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Skills4Attendance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------- Dashboard ----------
...
  /attendance/sessions:
    get:
      operationId: listAttendanceSessions
      tags: [attendance]
      parameters:
        - name: cohortId
          in: query
          schema: { type: integer }
        - name: tutorId
          in: query
          schema: { type: integer }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Attendance sessions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AttendanceSession" }
    post:
      operationId: createAttendanceSession
      tags: [attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceSessionInput"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSession"
        "409":
          description: Duplicate session exists, confirm to override
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/sessions/{id}:
    get:
      operationId: getAttendanceSession
      tags: [attendance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Attendance register for a session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRegister"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAttendanceSession
      tags: [attendance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceSessionUpdate"
      responses:
        "200":
          description: Session updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSession"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/sessions/{id}/register:
    put:
      operationId: saveAttendanceRegister
      tags: [attendance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceRegisterInput"
      responses:
        "200":
          description: Register saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRegister"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/sessions/{id}/mark-all-present:
    post:
      operationId: markAllPresent
      tags: [attendance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: All learners marked present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRegister"

  # ---------- Reports ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    UserRole:
      type: string
      enum: [admin, tutor]
...
    AttendanceStatus:
      type: string
      enum:
        - present
        - absent_authorised
        - absent_unauthorised
        - late
        - not_expected
        - withdrawn
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
    AuthUser:
      type: object
      required: [id, firstName, lastName, email, role]
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        tutorId: { type: ["integer", "null"] }
...
    AttendanceSession:
      type: object
      required:
        [
          id,
          cohortId,
          cohortName,
          tutorId,
          tutorName,
          sessionDate,
          plannedStartTime,
          plannedEndTime,
          plannedDurationHours,
          title,
          notes,
          recordedCount,
          expectedCount,
          createdBy,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        cohortId: { type: integer }
        cohortName: { type: string }
        tutorId: { type: integer }
        tutorName: { type: string }
        sessionDate: { type: string, format: date }
        plannedStartTime: { type: string }
        plannedEndTime: { type: string }
        plannedDurationHours: { type: number }
        title: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        recordedCount: { type: integer }
        expectedCount: { type: integer }
        createdBy: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AttendanceSessionInput:
      type: object
      required:
        [
          cohortId,
          sessionDate,
          plannedStartTime,
          plannedEndTime,
          plannedDurationHours,
        ]
      properties:
        cohortId: { type: integer }
        sessionDate: { type: string, format: date }
        plannedStartTime: { type: string, minLength: 1 }
        plannedEndTime: { type: string, minLength: 1 }
        plannedDurationHours: { type: number, minimum: 0 }
        title: { type: string }
        notes: { type: string }
        force: { type: boolean, default: false }
    AttendanceSessionUpdate:
      type: object
      properties:
        sessionDate: { type: string, format: date }
        plannedStartTime: { type: string, minLength: 1 }
        plannedEndTime: { type: string, minLength: 1 }
        plannedDurationHours: { type: number, minimum: 0 }
        title: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    RegisterEntry:
      type: object
      required:
        [
          learnerId,
          learnerName,
          learnerRef,
          status,
          hoursAttended,
          minutesLate,
          notes,
          overrideReason,
          recordId,
          lastEditedBy,
          lastEditedByName,
        ]
      properties:
        recordId: { type: ["integer", "null"] }
        learnerId: { type: integer }
        learnerName: { type: string }
        learnerRef: { type: string }
        status: { $ref: "#/components/schemas/AttendanceStatus" }
        hoursAttended: { type: number }
        minutesLate: { type: integer }
        notes: { type: ["string", "null"] }
        overrideReason: { type: ["string", "null"] }
        lastEditedBy: { type: ["integer", "null"] }
        lastEditedByName: { type: ["string", "null"] }
    RegisterEntryInput:
      type: object
      required: [learnerId, status, hoursAttended, minutesLate]
      properties:
        learnerId: { type: integer }
        status: { $ref: "#/components/schemas/AttendanceStatus" }
        hoursAttended: { type: number, minimum: 0 }
        minutesLate: { type: integer, minimum: 0 }
        notes: { type: string }
        overrideReason: { type: string }
    AttendanceRegisterInput:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/RegisterEntryInput" }
    AttendanceRegister:
      type: object
      required: [session, entries]
      properties:
        session: { $ref: "#/components/schemas/AttendanceSession" }
        entries:
          type: array
          items: { $ref: "#/components/schemas/RegisterEntry" }
...